Call store.ListIndexableRepos in (*repos).ListIndexable
The docstring for `(*repos).ListIndexable` says:

> ListIndexable calls database.IndexableRepos.List, with tracing. It lists
> ALL indexable repos which could include private user added repos.
> In addition, it only lists cloned repositories.

But it didn't call `database.IndexableRepos`.

So either the docstring was wrong or the code. This PR here makes the
code do what the docstring says, but I'm not sure whether that's
correct - looking for feedback here.

(Data point that speaks for this change: the `s.cache.List` method
called right before also calls `store.ListIndexableRepos` in case the
cache is invalid)